// src/net/bindings/types.ts
export const MyModuleRow = __t.object("MyModuleRow", {
  id: __t.u64(),
  slot: __t.u8(),
  targetKind: __t.u8(),
  targetId: __t.u32(),
  nextCycleAtMs: __t.u64(),
  stopAfterCycle: __t.bool(),
});
export type MyModuleRow = __Infer<typeof MyModuleRow>;
// ...
export const MyStatsRow = __t.object("MyStatsRow", {
  playerId: __t.u32(),
  honey: __t.u32(),
  cargo: __t.u16(),
  cargoGold: __t.u16(),
  energy: __t.f32(),
  energyAt: __t.f64(),
  hp: __t.u16(),
  lensLevel: __t.u8(),
  gatlingLevel: __t.u8(),
  quiverLevel: __t.u8(),
  brushLevel: __t.u8(),
  cargoLevel: __t.u8(),
  wingsLevel: __t.u8(),
  armorLevel: __t.u8(),
  antennaLevel: __t.u8(),
  nectarLevel: __t.u8(),
  dockedHive: __t.u8(),
  hivesVisited: __t.u16(),
  achievements: __t.u32(),
  deaths: __t.u32(),
  savedX: __t.f32(),
  savedY: __t.f32(),
  savedZ: __t.f32(),
  savedFlags: __t.u8(),
});
export type MyStatsRow = __Infer<typeof MyStatsRow>;